Static ordering of elements
@bgruening/@jmchilton, this might be useful if/when a replacement
normalize exists. This (pretty hacky) code will let order the elements
in a pleasant way for humans to read and review.

Unfortunatelly utils/gtree.py is mildly specific to parts of this case (i.e.
gwrite) because I couldn't get the standard etree.ElementTree.write()
function to call my overridden method. I'm sure I'm missing something
terribly simple...

If anyone has suggestions on how to get the inheritance working
properly such that calls to the parent .write() find the correct (i.e.
gtree) serialization function, then we can do a similar thing to
gxargparse and just shim this in front of etree!
