+++ tools/bcftools/bcftools_annotate.xml
@@ -1,5 +1,5 @@
-<?xml version="1.0"?>
-<tool id="bcftools_@EXECUTABLE@" name="bcftools @EXECUTABLE@" version="@VERSION@.0">
+<?xml version="1.0" encoding="utf-8"?>
+<tool name="bcftools @EXECUTABLE@" id="bcftools_@EXECUTABLE@" version="@VERSION@.0">
   <description>Annotate and edit VCF/BCF files.</description>
   <macros>
     <token name="@EXECUTABLE@">annotate</token>
@@ -64,37 +64,37 @@ $input_file
 &gt;
 $output_file</command>
   <inputs>
-    <param format="vcf,bcf" label="VCF/BCF Data" name="input_file" type="data"/>
-    <section expanded="true" name="sec_default" title="Default Options">
-      <param format="data" help="VCF file or tabix-indexed file with annotations: CHR\tPOS[\tVALUE]+" label="Annotations" name="annotations" optional="True" type="data"/>
+    <param name="input_file" label="VCF/BCF Data" type="data" format="vcf,bcf"/>
+    <section name="sec_default" expanded="true" title="Default Options">
+      <param name="annotations" label="Annotations" type="data" format="data" optional="True" help="VCF file or tabix-indexed file with annotations: CHR\tPOS[\tVALUE]+"/>
       <repeat name="columns_repeat" title="Columns">
-        <param help="list of columns in the annotation file, e.g. CHROM,POS,REF,ALT,-,INFO/TAG. See man page for details" label="Columns" name="columns"/>
+        <param name="columns" label="Columns" help="list of columns in the annotation file, e.g. CHROM,POS,REF,ALT,-,INFO/TAG. See man page for details"/>
       </repeat>
       <expand macro="macro_exclude"/>
-      <param format="data" help="lines which should be appended to the VCF header" label="Header Lines" name="header_lines" optional="True" type="data"/>
-      <param __TODO__="TODO" help="set ID column, see man pagee for details" label="Set Id" name="set_id"/>
-      <param __TODO__="TODO" help="select sites for which the expression is true (see man pagee for details)" label="Include" name="include"/>
+      <param name="header_lines" label="Header Lines" type="data" format="data" optional="True" help="lines which should be appended to the VCF header"/>
+      <param name="set_id" label="Set Id" __TODO__="TODO" help="set ID column, see man pagee for details"/>
+      <param name="include" label="Include" __TODO__="TODO" help="select sites for which the expression is true (see man pagee for details)"/>
       <expand macro="macro_output"/>
       <expand macro="macro_select_output_type"/>
       <expand macro="macro_regions_repeat"/>
       <expand macro="macro_regions_file"/>
-      <param format="data" help="rename sequences according to map file: from\tto" label="Rename Chrs" name="rename_chrs" optional="True" type="data"/>
+      <param name="rename_chrs" label="Rename Chrs" type="data" format="data" optional="True" help="rename sequences according to map file: from\tto"/>
       <expand macro="macro_samples_repeat"/>
       <expand macro="macro_invert_samples"/>
-      <param format="data" help="file of samples to annotate (or exclude with &quot;^&quot; prefix)" label="Samples File" name="samples_file" optional="True" type="data"/>
+      <param name="samples_file" label="Samples File" type="data" format="data" optional="True" help="file of samples to annotate (or exclude with &quot;^&quot; prefix)"/>
       <expand macro="macro_invert_samples_file"/>
       <repeat name="remove_repeat" title="Remove">
-        <param help="list of annotations to remove (e.g. ID,INFO/DP,FORMAT/DP,FILTER). See man page for details" label="Remove" name="remove"/>
+        <param name="remove" label="Remove" help="list of annotations to remove (e.g. ID,INFO/DP,FORMAT/DP,FILTER). See man page for details"/>
       </repeat>
     </section>
   </inputs>
   <outputs>
-    <data format="vcf" name="output_file">
+    <data name="output_file" format="vcf">
       <change_format>
-        <when format="bcf_bgz" input="sec_default|select_output_type" value="b"/>
-        <when format="bcf" input="sec_default|select_output_type" value="u"/>
-        <when format="vcf_bgz" input="sec_default|select_output_type" value="z"/>
-        <when format="vcf" input="sec_default|select_output_type" value="v"/>
+        <when input="sec_default|select_output_type" format="bcf_bgz" value="b"/>
+        <when input="sec_default|select_output_type" format="bcf" value="u"/>
+        <when input="sec_default|select_output_type" format="vcf_bgz" value="z"/>
+        <when input="sec_default|select_output_type" format="vcf" value="v"/>
       </change_format>
     </data>
   </outputs>
