+++ tools/bcftools/bcftools_concat.xml
@@ -1,5 +1,5 @@
-<?xml version="1.0"?>
-<tool id="bcftools_@EXECUTABLE@" name="bcftools @EXECUTABLE@" version="@VERSION@.0">
+<?xml version="1.0" encoding="utf-8"?>
+<tool name="bcftools @EXECUTABLE@" id="bcftools_@EXECUTABLE@" version="@VERSION@.0">
   <description>Concatenate or combine VCF/BCF files</description>
   <macros>
     <token name="@EXECUTABLE@">concat</token>
@@ -38,27 +38,27 @@ ${sec_default.ligate}
 
 </command>
   <inputs>
-    <param format="vcf,bcf" label="VCF/BCF Data" name="input_file1" type="data"/>
-    <param format="vcf,bcf" label="Other VCF/BCF Datasets" multiple="True" name="input_file2" type="data"/>
-    <section expanded="true" name="sec_default" title="Default Options">
-      <param falsevalue="" help="First coordinate of the next file can precede last record of the current file." label="Allow Overlaps" name="allow_overlaps" truevalue="--allow-overlaps" type="boolean"/>
-      <param falsevalue="" help="Output only once records present in multiple files." label="Remove Duplicates" name="remove_duplicates" truevalue="--remove-duplicates" type="boolean"/>
-      <param format="data" help="Read the list of files from a file." label="File List" name="file_list" optional="True" type="data"/>
-      <param falsevalue="" help="Ligate phased VCFs by matching phase at overlapping haplotypes" label="Ligate" name="ligate" truevalue="--ligate" type="boolean"/>
-      <param default="30" help="Break phase set if phasing quality is lower than &lt;int&gt;" label="Min Pq" name="min_PQ" optional="True" type="integer"/>
-      <param format="data" help="Write output to a file [standard output]" label="Output" name="output" optional="True" type="data"/>
+    <param name="input_file1" label="VCF/BCF Data" type="data" format="vcf,bcf"/>
+    <param name="input_file2" label="Other VCF/BCF Datasets" type="data" format="vcf,bcf" multiple="True"/>
+    <section name="sec_default" expanded="true" title="Default Options">
+      <param name="allow_overlaps" label="Allow Overlaps" type="boolean" truevalue="--allow-overlaps" falsevalue="" help="First coordinate of the next file can precede last record of the current file."/>
+      <param name="remove_duplicates" label="Remove Duplicates" type="boolean" truevalue="--remove-duplicates" falsevalue="" help="Output only once records present in multiple files."/>
+      <param name="file_list" label="File List" type="data" format="data" optional="True" help="Read the list of files from a file."/>
+      <param name="ligate" label="Ligate" type="boolean" truevalue="--ligate" falsevalue="" help="Ligate phased VCFs by matching phase at overlapping haplotypes"/>
+      <param name="min_PQ" label="Min Pq" type="integer" default="30" optional="True" help="Break phase set if phasing quality is lower than &lt;int&gt;"/>
+      <param name="output" label="Output" type="data" format="data" optional="True" help="Write output to a file [standard output]"/>
       <expand macro="macro_select_output_type"/>
       <expand macro="macro_regions_repeat"/>
       <expand macro="macro_regions_file"/>
     </section>
   </inputs>
   <outputs>
-    <data format="vcf" name="output_file">
+    <data name="output_file" format="vcf">
       <change_format>
-        <when format="bcf_bgz" input="sec_default|select_output_type" value="b"/>
-        <when format="bcf" input="sec_default|select_output_type" value="u"/>
-        <when format="vcf_bgz" input="sec_default|select_output_type" value="z"/>
-        <when format="vcf" input="sec_default|select_output_type" value="v"/>
+        <when input="sec_default|select_output_type" format="bcf_bgz" value="b"/>
+        <when input="sec_default|select_output_type" format="bcf" value="u"/>
+        <when input="sec_default|select_output_type" format="vcf_bgz" value="z"/>
+        <when input="sec_default|select_output_type" format="vcf" value="v"/>
       </change_format>
     </data>
   </outputs>
