+++ tools/bcftools/bcftools_consensus.xml
@@ -1,5 +1,5 @@
-<?xml version="1.0"?>
-<tool id="bcftools_@EXECUTABLE@" name="bcftools @EXECUTABLE@" version="@VERSION@.0">
+<?xml version="1.0" encoding="utf-8"?>
+<tool name="bcftools @EXECUTABLE@" id="bcftools_@EXECUTABLE@" version="@VERSION@.0">
   <description>Create consensus sequence by applying VCF variants to a reference fasta file.</description>
   <macros>
     <token name="@EXECUTABLE@">consensus</token>
@@ -37,23 +37,23 @@ $input_file
 &gt;
 $output_file</command>
   <inputs>
-    <param format="fasta" label="Reference Fasta" name="reference_fasta" type="data"/>
-    <param format="vcf_bgz,bcf_bgz" label="VCF/BCF Data" name="input_file" type="data"/>
-    <section expanded="true" name="sec_default" title="Default Options">
-      <param format="data" help="reference sequence in fasta format" label="Fasta Ref" name="fasta_ref" optional="True" type="data"/>
+    <param name="reference_fasta" label="Reference Fasta" type="data" format="fasta"/>
+    <param name="input_file" label="VCF/BCF Data" type="data" format="vcf_bgz,bcf_bgz"/>
+    <section name="sec_default" expanded="true" title="Default Options">
+      <param name="fasta_ref" label="Fasta Ref" type="data" format="data" optional="True" help="reference sequence in fasta format"/>
       <param name="select_haplotype" type="select">
-        <option selected="True" value="__none__">No selection</option>
+        <option value="__none__" selected="True">No selection</option>
         <option value="apply variants for the given haplotype"/>
       </param>
-      <param falsevalue="" help="output variants in the form of IUPAC ambiguity codes" label="Iupac Codes" name="iupac_codes" truevalue="--iupac-codes" type="boolean"/>
-      <param format="data" help="replace regions with N" label="Mask" name="mask" optional="True" type="data"/>
+      <param name="iupac_codes" label="Iupac Codes" type="boolean" truevalue="--iupac-codes" falsevalue="" help="output variants in the form of IUPAC ambiguity codes"/>
+      <param name="mask" label="Mask" type="data" format="data" optional="True" help="replace regions with N"/>
       <expand macro="macro_output"/>
-      <param format="data" help="write a chain file for liftover" label="Chain" name="chain" optional="True" type="data"/>
-      <param help="apply variants of the given sample" label="Sample" name="sample" optional="True" type="string"/>
+      <param name="chain" label="Chain" type="data" format="data" optional="True" help="write a chain file for liftover"/>
+      <param name="sample" label="Sample" type="string" optional="True" help="apply variants of the given sample"/>
     </section>
   </inputs>
   <outputs>
-    <data format="fasta" name="output_file"/>
+    <data name="output_file" format="fasta"/>
   </outputs>
   <tests/>
   <help>Create consensus sequence by applying VCF variants to a reference fasta file.</help>
