+++ tools/bcftools/bcftools_filter.xml
@@ -1,5 +1,5 @@
-<?xml version="1.0"?>
-<tool id="bcftools_@EXECUTABLE@" name="bcftools @EXECUTABLE@" version="@VERSION@.0">
+<?xml version="1.0" encoding="utf-8"?>
+<tool name="bcftools @EXECUTABLE@" id="bcftools_@EXECUTABLE@" version="@VERSION@.0">
   <description>Apply fixed-threshold filters.</description>
   <macros>
     <token name="@EXECUTABLE@">filter</token>
@@ -59,20 +59,20 @@ $input_file
 &gt;
 $output_file</command>
   <inputs>
-    <param format="vcf,bcf" label="VCF/BCF Data" name="input_file" type="data"/>
-    <section expanded="true" name="sec_default" title="Default Options">
+    <param name="input_file" label="VCF/BCF Data" type="data" format="vcf,bcf"/>
+    <section name="sec_default" expanded="true" title="Default Options">
       <expand macro="macro_exclude"/>
-      <param help="filter SNPs within &lt;int&gt; base pairs of an indel" label="Snpgap" name="SnpGap" optional="True" type="integer"/>
-      <param help="filter clusters of indels separated by &lt;int&gt; or fewer base pairs allowing only one to pass" label="Indelgap" name="IndelGap" optional="True" type="integer"/>
-      <param __TODO__="TODO" help="include only sites for which the expression is true (see man page for details" label="Include" name="include"/>
-      <param __TODO__="TODO" help="&quot;+&quot;: do not replace but add to existing FILTER; &quot;x&quot;: reset filters at sites which pass" label="Mode" name="mode"/>
+      <param name="SnpGap" label="Snpgap" type="integer" optional="True" help="filter SNPs within &lt;int&gt; base pairs of an indel"/>
+      <param name="IndelGap" label="Indelgap" type="integer" optional="True" help="filter clusters of indels separated by &lt;int&gt; or fewer base pairs allowing only one to pass"/>
+      <param name="include" label="Include" __TODO__="TODO" help="include only sites for which the expression is true (see man page for details"/>
+      <param name="mode" label="Mode" __TODO__="TODO" help="&quot;+&quot;: do not replace but add to existing FILTER; &quot;x&quot;: reset filters at sites which pass"/>
       <expand macro="macro_output"/>
       <expand macro="macro_select_output_type"/>
       <expand macro="macro_regions_repeat"/>
       <expand macro="macro_regions_file"/>
-      <param help="annotate FILTER column with &lt;string&gt; or unique filter name (&quot;Filter%d&quot;) made up by the program (&quot;+&quot;)" label="Soft Filter" name="soft_filter" optional="True" type="string"/>
+      <param name="soft_filter" label="Soft Filter" type="string" optional="True" help="annotate FILTER column with &lt;string&gt; or unique filter name (&quot;Filter%d&quot;) made up by the program (&quot;+&quot;)"/>
       <param name="select_set_GTs" type="select">
-        <option selected="True" value="__none__">No selection</option>
+        <option value="__none__" selected="True">No selection</option>
         <option value="set genotypes of failed samples to missing (.) or ref (0)"/>
       </param>
       <expand macro="macro_targets_repeat"/>
@@ -80,12 +80,12 @@ $output_file</command>
     </section>
   </inputs>
   <outputs>
-    <data format="vcf" name="output_file">
+    <data name="output_file" format="vcf">
       <change_format>
-        <when format="bcf_bgz" input="sec_default|select_output_type" value="b"/>
-        <when format="bcf" input="sec_default|select_output_type" value="u"/>
-        <when format="vcf_bgz" input="sec_default|select_output_type" value="z"/>
-        <when format="vcf" input="sec_default|select_output_type" value="v"/>
+        <when input="sec_default|select_output_type" format="bcf_bgz" value="b"/>
+        <when input="sec_default|select_output_type" format="bcf" value="u"/>
+        <when input="sec_default|select_output_type" format="vcf_bgz" value="z"/>
+        <when input="sec_default|select_output_type" format="vcf" value="v"/>
       </change_format>
     </data>
   </outputs>
