+++ tools/bcftools/bcftools_gtcheck.xml
@@ -1,5 +1,5 @@
-<?xml version="1.0"?>
-<tool id="bcftools_@EXECUTABLE@" name="bcftools @EXECUTABLE@" version="@VERSION@.0">
+<?xml version="1.0" encoding="utf-8"?>
+<tool name="bcftools @EXECUTABLE@" id="bcftools_@EXECUTABLE@" version="@VERSION@.0">
   <description>Check sample identity</description>
   <macros>
     <token name="@EXECUTABLE@">gtcheck</token>
@@ -52,24 +52,24 @@ $query_file
 &gt;
 $output_file</command>
   <inputs>
-    <param format="vcf,bcf" label="Genotypes VCF/BCF Data" name="genotypes_file" optional="True" type="data"/>
-    <param format="vcf,bcf" label="Query VCF/BCF Data" name="query_file" type="data"/>
-    <section expanded="true" name="sec_default" title="Default Options">
-      <param falsevalue="" help="output comparison for all sites" label="All Sites" name="all_sites" truevalue="--all-sites" type="boolean"/>
-      <param format="data" help="genotypes to compare against" label="Genotypes" name="genotypes" optional="True" type="data"/>
-      <param default="99" help="use GTs, ignore PLs, using &lt;int&gt; for unseen genotypes" label="Gts Only" name="GTs_only" optional="True" type="integer"/>
-      <param falsevalue="" help="homozygous genotypes only (useful for low coverage data)" label="Homs Only" name="homs_only" truevalue="--homs-only" type="boolean"/>
-      <param __TODO__="TODO" help="plot" label="Plot" name="plot"/>
+    <param name="genotypes_file" label="Genotypes VCF/BCF Data" type="data" format="vcf,bcf" optional="True"/>
+    <param name="query_file" label="Query VCF/BCF Data" type="data" format="vcf,bcf"/>
+    <section name="sec_default" expanded="true" title="Default Options">
+      <param name="all_sites" label="All Sites" type="boolean" truevalue="--all-sites" falsevalue="" help="output comparison for all sites"/>
+      <param name="genotypes" label="Genotypes" type="data" format="data" optional="True" help="genotypes to compare against"/>
+      <param name="GTs_only" label="Gts Only" type="integer" default="99" optional="True" help="use GTs, ignore PLs, using &lt;int&gt; for unseen genotypes"/>
+      <param name="homs_only" label="Homs Only" type="boolean" truevalue="--homs-only" falsevalue="" help="homozygous genotypes only (useful for low coverage data)"/>
+      <param name="plot" label="Plot" __TODO__="TODO" help="plot"/>
       <expand macro="macro_regions_repeat"/>
       <expand macro="macro_regions_file"/>
-      <param help="query sample (by default the first sample is checked)" label="Query Sample" name="query_sample" optional="True" type="string"/>
-      <param help="target sample in the -g file (used only for plotting)" label="Target Sample" name="target_sample" optional="True" type="string"/>
+      <param name="query_sample" label="Query Sample" type="string" optional="True" help="query sample (by default the first sample is checked)"/>
+      <param name="target_sample" label="Target Sample" type="string" optional="True" help="target sample in the -g file (used only for plotting)"/>
       <expand macro="macro_targets_repeat"/>
       <expand macro="macro_targets_file"/>
     </section>
   </inputs>
   <outputs>
-    <data format="tabular" name="output_file"/>
+    <data name="output_file" format="tabular"/>
   </outputs>
   <tests/>
   <help>Check sample identity. With no -g BCF given, multi-sample cross-check is performed.</help>
