+++ tools/bcftools/bcftools_isec.xml
@@ -1,5 +1,5 @@
-<?xml version="1.0"?>
-<tool id="bcftools_@EXECUTABLE@" name="bcftools @EXECUTABLE@" version="@VERSION@.0">
+<?xml version="1.0" encoding="utf-8"?>
+<tool name="bcftools @EXECUTABLE@" id="bcftools_@EXECUTABLE@" version="@VERSION@.0">
   <description>Create intersections, unions and complements of VCF files.</description>
   <macros>
     <token name="@EXECUTABLE@">isec</token>
@@ -60,34 +60,34 @@ ${sec_default.complement}
 
 </command>
   <inputs>
-    <param format="vcf,bcf" label="VCF/BCF Data" name="input_file1" type="data"/>
-    <param format="vcf,bcf" label="Other VCF/BCF Datasets" multiple="True" name="input_file2" type="data"/>
-    <section expanded="true" name="sec_default" title="Default Options">
+    <param name="input_file1" label="VCF/BCF Data" type="data" format="vcf,bcf"/>
+    <param name="input_file2" label="Other VCF/BCF Datasets" type="data" format="vcf,bcf" multiple="True"/>
+    <section name="sec_default" expanded="true" title="Default Options">
       <expand macro="macro_collapse"/>
-      <param falsevalue="" help="output positions present only in the first file but missing in the others" label="Complement" name="complement" truevalue="--complement" type="boolean"/>
-      <param __TODO__="TODO" help="exclude sites for which the expression is true" label="Exclude" name="exclude"/>
+      <param name="complement" label="Complement" type="boolean" truevalue="--complement" falsevalue="" help="output positions present only in the first file but missing in the others"/>
+      <param name="exclude" label="Exclude" __TODO__="TODO" help="exclude sites for which the expression is true"/>
       <expand macro="macro_apply_filters_repeat"/>
-      <param __TODO__="TODO" help="include only sites for which the expression is true" label="Include" name="include"/>
-      <param help="output positions present in this many (=), this many or more (+), or this many or fewer (-) files" label="Nfiles" name="nfiles" optional="True" type="integer"/>
+      <param name="include" label="Include" __TODO__="TODO" help="include only sites for which the expression is true"/>
+      <param name="nfiles" label="Nfiles" type="integer" optional="True" help="output positions present in this many (=), this many or more (+), or this many or fewer (-) files"/>
       <expand macro="macro_output"/>
       <expand macro="macro_select_output_type"/>
-      <param __TODO__="TODO" help="if given, subset each of the input files accordingly, see also -w" label="Prefix" name="prefix"/>
+      <param name="prefix" label="Prefix" __TODO__="TODO" help="if given, subset each of the input files accordingly, see also -w"/>
       <expand macro="macro_regions_repeat"/>
       <expand macro="macro_regions_file"/>
       <expand macro="macro_targets_repeat"/>
       <expand macro="macro_targets_file"/>
       <repeat name="write_repeat" title="Write">
-        <param help="list of files to write with -p given as 1-based indexes. By default, all files are written" label="Write" name="write"/>
+        <param name="write" label="Write" help="list of files to write with -p given as 1-based indexes. By default, all files are written"/>
       </repeat>
     </section>
   </inputs>
   <outputs>
-    <data format="vcf" name="output_file">
+    <data name="output_file" format="vcf">
       <change_format>
-        <when format="bcf_bgz" input="sec_default|select_output_type" value="b"/>
-        <when format="bcf" input="sec_default|select_output_type" value="u"/>
-        <when format="vcf_bgz" input="sec_default|select_output_type" value="z"/>
-        <when format="vcf" input="sec_default|select_output_type" value="v"/>
+        <when input="sec_default|select_output_type" format="bcf_bgz" value="b"/>
+        <when input="sec_default|select_output_type" format="bcf" value="u"/>
+        <when input="sec_default|select_output_type" format="vcf_bgz" value="z"/>
+        <when input="sec_default|select_output_type" format="vcf" value="v"/>
       </change_format>
     </data>
   </outputs>
