+++ tools/bcftools/bcftools_merge.xml
@@ -1,5 +1,5 @@
-<?xml version="1.0"?>
-<tool id="bcftools_@EXECUTABLE@" name="bcftools @EXECUTABLE@" version="@VERSION@.0">
+<?xml version="1.0" encoding="utf-8"?>
+<tool name="bcftools @EXECUTABLE@" id="bcftools_@EXECUTABLE@" version="@VERSION@.0">
   <description>Merge multiple VCF/BCF files from non-overlapping sample sets to create one multi-sample file</description>
   <macros>
     <token name="@EXECUTABLE@">merge</token>
@@ -47,16 +47,16 @@ ${sec_default.print_header}
 
 </command>
   <inputs>
-    <param format="vcf,bcf" label="VCF/BCF Data" name="input_file1" type="data"/>
-    <param format="vcf,bcf" label="Other VCF/BCF Datasets" multiple="True" name="input_file2" type="data"/>
-    <section expanded="true" name="sec_default" title="Default Options">
-      <param falsevalue="" help="resolve duplicate sample names" label="Force Samples" name="force_samples" truevalue="--force-samples" type="boolean"/>
-      <param falsevalue="" help="print only the merged header and exit" label="Print Header" name="print_header" truevalue="--print-header" type="boolean"/>
-      <param format="data" help="use the provided header" label="Use Header" name="use_header" optional="True" type="data"/>
+    <param name="input_file1" label="VCF/BCF Data" type="data" format="vcf,bcf"/>
+    <param name="input_file2" label="Other VCF/BCF Datasets" type="data" format="vcf,bcf" multiple="True"/>
+    <section name="sec_default" expanded="true" title="Default Options">
+      <param name="force_samples" label="Force Samples" type="boolean" truevalue="--force-samples" falsevalue="" help="resolve duplicate sample names"/>
+      <param name="print_header" label="Print Header" type="boolean" truevalue="--print-header" falsevalue="" help="print only the merged header and exit"/>
+      <param name="use_header" label="Use Header" type="data" format="data" optional="True" help="use the provided header"/>
       <expand macro="macro_apply_filters_repeat"/>
-      <param __TODO__="TODO" help="rules for merging INFO fields (method is one of sum,avg,min,max,join) or &quot;-&quot; to turn off the default [DP:sum,DP4:sum]" label="Info Rules" name="info_rules"/>
-      <param format="data" help="read file names from the file" label="File List" name="file_list" optional="True" type="data"/>
-      <param help="allow multiallelic records for &lt;snps|indels|both|all|none|id&gt;, see man page for details [both]" label="Merge" name="merge" optional="True" type="string"/>
+      <param name="info_rules" label="Info Rules" __TODO__="TODO" help="rules for merging INFO fields (method is one of sum,avg,min,max,join) or &quot;-&quot; to turn off the default [DP:sum,DP4:sum]"/>
+      <param name="file_list" label="File List" type="data" format="data" optional="True" help="read file names from the file"/>
+      <param name="merge" label="Merge" type="string" optional="True" help="allow multiallelic records for &lt;snps|indels|both|all|none|id&gt;, see man page for details [both]"/>
       <expand macro="macro_output"/>
       <expand macro="macro_select_output_type"/>
       <expand macro="macro_regions_repeat"/>
@@ -64,12 +64,12 @@ ${sec_default.print_header}
     </section>
   </inputs>
   <outputs>
-    <data format="vcf" name="output_file">
+    <data name="output_file" format="vcf">
       <change_format>
-        <when format="bcf_bgz" input="sec_default|select_output_type" value="b"/>
-        <when format="bcf" input="sec_default|select_output_type" value="u"/>
-        <when format="vcf_bgz" input="sec_default|select_output_type" value="z"/>
-        <when format="vcf" input="sec_default|select_output_type" value="v"/>
+        <when input="sec_default|select_output_type" format="bcf_bgz" value="b"/>
+        <when input="sec_default|select_output_type" format="bcf" value="u"/>
+        <when input="sec_default|select_output_type" format="vcf_bgz" value="z"/>
+        <when input="sec_default|select_output_type" format="vcf" value="v"/>
       </change_format>
     </data>
   </outputs>
