+++ tools/bcftools/bcftools_norm.xml
@@ -1,5 +1,5 @@
-<?xml version="1.0"?>
-<tool id="bcftools_@EXECUTABLE@" name="bcftools @EXECUTABLE@" version="@VERSION@.0">
+<?xml version="1.0" encoding="utf-8"?>
+<tool name="bcftools @EXECUTABLE@" id="bcftools_@EXECUTABLE@" version="@VERSION@.0">
   <description>Left-align and normalize indels; check if REF alleles match the reference; split multiallelic sites into multiple rows; recover multiallelics from multiple rows.</description>
   <macros>
     <token name="@EXECUTABLE@">norm</token>
@@ -52,15 +52,15 @@ $input_file
 &gt;
 $output_file</command>
   <inputs>
-    <param format="vcf,bcf" label="VCF/BCF Data" name="input_file" type="data"/>
-    <section expanded="true" name="sec_default" title="Default Options">
+    <param name="input_file" label="VCF/BCF Data" type="data" format="vcf,bcf"/>
+    <section name="sec_default" expanded="true" title="Default Options">
       <param name="select_check_ref" type="select">
         <option value="check REF alleles and exit (e)"/>
         <option value="warn (w)"/>
         <option value="exclude (x) bad sites"/>
       </param>
-      <param falsevalue="" help="remove duplicate lines of the same type." label="Remove Duplicates" name="remove_duplicates" truevalue="--remove-duplicates" type="boolean"/>
-      <param format="data" help="reference sequence" label="Fasta Ref" name="fasta_ref" optional="True" type="data"/>
+      <param name="remove_duplicates" label="Remove Duplicates" type="boolean" truevalue="--remove-duplicates" falsevalue="" help="remove duplicate lines of the same type."/>
+      <param name="fasta_ref" label="Fasta Ref" type="data" format="data" optional="True" help="reference sequence"/>
       <param name="select_multiallelics" type="select">
         <option value="split multiallelics (-) or join biallelics (+)"/>
         <option value="type">snps|indels|both|any</option>
@@ -69,19 +69,19 @@ $output_file</command>
       <expand macro="macro_select_output_type"/>
       <expand macro="macro_regions_repeat"/>
       <expand macro="macro_regions_file"/>
-      <param falsevalue="" help="when merging (-m+), merged site is PASS only if all sites being merged PASS" label="Strict Filter" name="strict_filter" truevalue="--strict-filter" type="boolean"/>
+      <param name="strict_filter" label="Strict Filter" type="boolean" truevalue="--strict-filter" falsevalue="" help="when merging (-m+), merged site is PASS only if all sites being merged PASS"/>
       <expand macro="macro_targets_repeat"/>
       <expand macro="macro_targets_file"/>
-      <param default="1000" help="buffer for sorting lines which changed position during realignment" label="Site Win" name="site_win" optional="True" type="integer"/>
+      <param name="site_win" label="Site Win" type="integer" default="1000" optional="True" help="buffer for sorting lines which changed position during realignment"/>
     </section>
   </inputs>
   <outputs>
-    <data format="vcf" name="output_file">
+    <data name="output_file" format="vcf">
       <change_format>
-        <when format="bcf_bgz" input="sec_default|select_output_type" value="b"/>
-        <when format="bcf" input="sec_default|select_output_type" value="u"/>
-        <when format="vcf_bgz" input="sec_default|select_output_type" value="z"/>
-        <when format="vcf" input="sec_default|select_output_type" value="v"/>
+        <when input="sec_default|select_output_type" format="bcf_bgz" value="b"/>
+        <when input="sec_default|select_output_type" format="bcf" value="u"/>
+        <when input="sec_default|select_output_type" format="vcf_bgz" value="z"/>
+        <when input="sec_default|select_output_type" format="vcf" value="v"/>
       </change_format>
     </data>
   </outputs>
