+++ tools/bcftools/bcftools_query.xml
@@ -1,5 +1,5 @@
-<?xml version="1.0"?>
-<tool id="bcftools_@EXECUTABLE@" name="bcftools @EXECUTABLE@" version="@VERSION@.0">
+<?xml version="1.0" encoding="utf-8"?>
+<tool name="bcftools @EXECUTABLE@" id="bcftools_@EXECUTABLE@" version="@VERSION@.0">
   <description>Extracts fields from VCF/BCF file and prints them in user-defined format</description>
   <macros>
     <token name="@EXECUTABLE@">query</token>
@@ -57,15 +57,15 @@ ${sec_default.list_samples}
 
 </command>
   <inputs>
-    <param format="vcf,bcf" label="VCF/BCF Data" name="input_file1" type="data"/>
-    <param format="vcf,bcf" label="Other VCF/BCF Datasets" multiple="True" name="input_file2" type="data"/>
-    <section expanded="true" name="sec_default" title="Default Options">
-      <param help="collapse lines with duplicate positions for &lt;snps|indels|both|all|some|none&gt;, see man page [none]" label="Collapse" name="collapse" optional="True" type="string"/>
+    <param name="input_file1" label="VCF/BCF Data" type="data" format="vcf,bcf"/>
+    <param name="input_file2" label="Other VCF/BCF Datasets" type="data" format="vcf,bcf" multiple="True"/>
+    <section name="sec_default" expanded="true" title="Default Options">
+      <param name="collapse" label="Collapse" type="string" optional="True" help="collapse lines with duplicate positions for &lt;snps|indels|both|all|some|none&gt;, see man page [none]"/>
       <expand macro="macro_exclude"/>
-      <param help="see man page for details" label="Format" name="format" optional="True" type="string"/>
-      <param falsevalue="" help="print header" label="Print Header" name="print_header" truevalue="--print-header" type="boolean"/>
-      <param __TODO__="TODO" help="select sites for which the expression is true (see man page for details)" label="Include" name="include"/>
-      <param falsevalue="" help="print the list of samples and exit" label="List Samples" name="list_samples" truevalue="--list-samples" type="boolean"/>
+      <param name="format" label="Format" type="string" optional="True" help="see man page for details"/>
+      <param name="print_header" label="Print Header" type="boolean" truevalue="--print-header" falsevalue="" help="print header"/>
+      <param name="include" label="Include" __TODO__="TODO" help="select sites for which the expression is true (see man page for details)"/>
+      <param name="list_samples" label="List Samples" type="boolean" truevalue="--list-samples" falsevalue="" help="print the list of samples and exit"/>
       <expand macro="macro_output_file"/>
       <expand macro="macro_regions_repeat"/>
       <expand macro="macro_regions_file"/>
@@ -73,7 +73,7 @@ ${sec_default.list_samples}
       <expand macro="macro_samples_file"/>
       <expand macro="macro_targets_repeat"/>
       <expand macro="macro_targets_file"/>
-      <param format="data" help="process multiple VCFs listed in the file" label="Vcf List" name="vcf_list" optional="True" type="data"/>
+      <param name="vcf_list" label="Vcf List" type="data" format="data" optional="True" help="process multiple VCFs listed in the file"/>
     </section>
   </inputs>
   <outputs/>
