+++ tools/bcftools/bcftools_reheader.xml
@@ -1,5 +1,5 @@
-<?xml version="1.0"?>
-<tool id="bcftools_@EXECUTABLE@" name="bcftools @EXECUTABLE@" version="@VERSION@.0">
+<?xml version="1.0" encoding="utf-8"?>
+<tool name="bcftools @EXECUTABLE@" id="bcftools_@EXECUTABLE@" version="@VERSION@.0">
   <description>Modify header of VCF/BCF files, change sample names.</description>
   <macros>
     <token name="@EXECUTABLE@">reheader</token>
@@ -27,11 +27,11 @@ $input_file
 &gt;
 $output_file</command>
   <inputs>
-    <param format="vcf,bcf" label="VCF/BCF Data" name="input_file" type="data"/>
-    <section expanded="true" name="sec_default" title="Default Options">
-      <param format="data" help="new header" label="Header" name="header" optional="True" type="data"/>
+    <param name="input_file" label="VCF/BCF Data" type="data" format="vcf,bcf"/>
+    <section name="sec_default" expanded="true" title="Default Options">
+      <param name="header" label="Header" type="data" format="data" optional="True" help="new header"/>
       <expand macro="macro_output"/>
-      <param format="data" help="new sample names" label="Samples" name="samples" optional="True" type="data"/>
+      <param name="samples" label="Samples" type="data" format="data" optional="True" help="new sample names"/>
     </section>
   </inputs>
   <outputs/>
