+++ tools/bcftools/bcftools_roh.xml
@@ -1,5 +1,5 @@
-<?xml version="1.0"?>
-<tool id="bcftools_@EXECUTABLE@" name="bcftools @EXECUTABLE@" version="@VERSION@.0">
+<?xml version="1.0" encoding="utf-8"?>
+<tool name="bcftools @EXECUTABLE@" id="bcftools_@EXECUTABLE@" version="@VERSION@.0">
   <description>HMM model for detecting runs of autozygosity.</description>
   <macros>
     <token name="@EXECUTABLE@">roh</token>
@@ -63,25 +63,25 @@ $input_file
 &gt;
 $output_file</command>
   <inputs>
-    <param format="vcf,bcf" label="VCF/BCF Data" name="input_file" type="data"/>
-    <section expanded="true" name="sec_general" title="General Options">
-      <param __TODO__="TODO" help="use TAG for allele frequency" label="Af Tag" name="AF_tag"/>
-      <param format="data" help="read allele frequencies from file (CHR\tPOS\tREF,ALT\tAF)" label="Af File" name="AF_file" optional="True" type="data"/>
-      <param format="data" help="calculate AC,AN counts on the fly, using either all samples (&quot;-&quot;) or samples listed in &lt;file&gt;" label="Estimate Af" name="estimate_AF" optional="True" type="data"/>
-      <param help="use GTs, ignore PLs, use &lt;float&gt; for PL of unseen genotypes. Safe value to use is 30 to account for GT errors." label="Gts Only" name="GTs_only" optional="True" type="float"/>
-      <param falsevalue="" help="skip indels as their genotypes are enriched for errors" label="Skip Indels" name="skip_indels" truevalue="--skip-indels" type="boolean"/>
-      <param format="data" help="genetic map in IMPUTE2 format, single file or mask, where string &quot;{CHROM}&quot; is replaced with chromosome name" label="Genetic Map" name="genetic_map" optional="True" type="data"/>
-      <param help="constant recombination rate per bp" label="Rec Rate" name="rec_rate" optional="True" type="float"/>
+    <param name="input_file" label="VCF/BCF Data" type="data" format="vcf,bcf"/>
+    <section name="sec_general" expanded="true" title="General Options">
+      <param name="AF_tag" label="Af Tag" __TODO__="TODO" help="use TAG for allele frequency"/>
+      <param name="AF_file" label="Af File" type="data" format="data" optional="True" help="read allele frequencies from file (CHR\tPOS\tREF,ALT\tAF)"/>
+      <param name="estimate_AF" label="Estimate Af" type="data" format="data" optional="True" help="calculate AC,AN counts on the fly, using either all samples (&quot;-&quot;) or samples listed in &lt;file&gt;"/>
+      <param name="GTs_only" label="Gts Only" type="float" optional="True" help="use GTs, ignore PLs, use &lt;float&gt; for PL of unseen genotypes. Safe value to use is 30 to account for GT errors."/>
+      <param name="skip_indels" label="Skip Indels" type="boolean" truevalue="--skip-indels" falsevalue="" help="skip indels as their genotypes are enriched for errors"/>
+      <param name="genetic_map" label="Genetic Map" type="data" format="data" optional="True" help="genetic map in IMPUTE2 format, single file or mask, where string &quot;{CHROM}&quot; is replaced with chromosome name"/>
+      <param name="rec_rate" label="Rec Rate" type="float" optional="True" help="constant recombination rate per bp"/>
       <expand macro="macro_regions_repeat"/>
       <expand macro="macro_regions_file"/>
-      <param __TODO__="TODO" help="sample to analyze" label="Sample" name="sample"/>
+      <param name="sample" label="Sample" __TODO__="TODO" help="sample to analyze"/>
       <expand macro="macro_targets_repeat"/>
       <expand macro="macro_targets_file"/>
     </section>
-    <section expanded="true" name="sec_hmm" title="HMM Options">
-      <param default="1e-8" help="P(AZ|HW) transition probability from AZ (autozygous) to HW (Hardy-Weinberg) state" label="Hw To Az" name="hw_to_az" optional="True" type="float"/>
-      <param default="1e-7" help="P(HW|AZ) transition probability from HW to AZ state" label="Az To Hw" name="az_to_hw" optional="True" type="float"/>
-      <param falsevalue="" help="perform Viterbi training to estimate transition probabilities" label="Viterbi Training" name="viterbi_training" truevalue="--viterbi-training" type="boolean"/>
+    <section name="sec_hmm" expanded="true" title="HMM Options">
+      <param name="hw_to_az" label="Hw To Az" type="float" default="1e-8" optional="True" help="P(AZ|HW) transition probability from AZ (autozygous) to HW (Hardy-Weinberg) state"/>
+      <param name="az_to_hw" label="Az To Hw" type="float" default="1e-7" optional="True" help="P(HW|AZ) transition probability from HW to AZ state"/>
+      <param name="viterbi_training" label="Viterbi Training" type="boolean" truevalue="--viterbi-training" falsevalue="" help="perform Viterbi training to estimate transition probabilities"/>
     </section>
   </inputs>
   <outputs/>
