+++ tools/bcftools/bcftools_stats.xml
@@ -1,5 +1,5 @@
-<?xml version="1.0"?>
-<tool id="bcftools_@EXECUTABLE@" name="bcftools @EXECUTABLE@" version="@VERSION@.0">
+<?xml version="1.0" encoding="utf-8"?>
+<tool name="bcftools @EXECUTABLE@" id="bcftools_@EXECUTABLE@" version="@VERSION@.0">
   <description>Parses VCF or BCF and produces stats which can be plotted using plot-vcfstats</description>
   <macros>
     <token name="@EXECUTABLE@">stats</token>
@@ -59,24 +59,24 @@ ${sec_default.verbose}
 
 </command>
   <inputs>
-    <param format="vcf,bcf" label="VCF/BCF Data" name="input_file1" type="data"/>
-    <param format="vcf,bcf" label="Second VCF/BCF Data" name="input_file2" optional="True" type="data"/>
-    <section expanded="true" name="sec_default" title="Default Options">
-      <param falsevalue="" help="include only 1st allele at multiallelic sites" label="1St Allele Only" name="1st_allele_only" truevalue="--1st-allele-only" type="boolean"/>
+    <param name="input_file1" label="VCF/BCF Data" type="data" format="vcf,bcf"/>
+    <param name="input_file2" label="Second VCF/BCF Data" type="data" format="vcf,bcf" optional="True"/>
+    <section name="sec_default" expanded="true" title="Default Options">
+      <param name="1st_allele_only" label="1St Allele Only" type="boolean" truevalue="--1st-allele-only" falsevalue="" help="include only 1st allele at multiallelic sites"/>
       <expand macro="macro_collapse"/>
-      <param __TODO__="TODO" help="depth distribution: min,max,bin size [0,500,1]" label="Depth" name="depth"/>
-      <param __TODO__="TODO" help="tab-delimited file with exons for indel frameshifts (chr,from,to; 1-based, inclusive, bgzip compressed)" label="Exons" name="exons"/>
+      <param name="depth" label="Depth" __TODO__="TODO" help="depth distribution: min,max,bin size [0,500,1]"/>
+      <param name="exons" label="Exons" __TODO__="TODO" help="tab-delimited file with exons for indel frameshifts (chr,from,to; 1-based, inclusive, bgzip compressed)"/>
       <expand macro="macro_apply_filters_repeat"/>
-      <param format="data" help="faidx indexed reference sequence file to determine INDEL context" label="Fasta Ref" name="fasta_ref" optional="True" type="data"/>
-      <param falsevalue="" help="collect stats for sites with ID separately (known vs novel)" label="Split By Id" name="split_by_ID" truevalue="--split-by-ID" type="boolean"/>
+      <param name="fasta_ref" label="Fasta Ref" type="data" format="data" optional="True" help="faidx indexed reference sequence file to determine INDEL context"/>
+      <param name="split_by_ID" label="Split By Id" type="boolean" truevalue="--split-by-ID" falsevalue="" help="collect stats for sites with ID separately (known vs novel)"/>
       <expand macro="macro_regions_repeat"/>
       <expand macro="macro_regions_file"/>
       <expand macro="macro_samples_repeat"/>
       <expand macro="macro_samples_file"/>
       <expand macro="macro_targets_repeat"/>
       <expand macro="macro_targets_file"/>
-      <param __TODO__="TODO" help="collect Ts/Tv stats for any tag using the given binning [0:1:100]" label="User Tstv" name="user_tstv"/>
-      <param falsevalue="" help="produce verbose per-site and per-sample output" label="Verbose" name="verbose" truevalue="--verbose" type="boolean"/>
+      <param name="user_tstv" label="User Tstv" __TODO__="TODO" help="collect Ts/Tv stats for any tag using the given binning [0:1:100]"/>
+      <param name="verbose" label="Verbose" type="boolean" truevalue="--verbose" falsevalue="" help="produce verbose per-site and per-sample output"/>
     </section>
   </inputs>
   <outputs/>
